requirements.txt
google-api-python-client==2.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
gspread==5.1.1